Added capsules, adjusted behaviors.
Friendly and enemy capsules are now tracked.

Defender will now chase enemies even if they are far away (->seen by the attacker)

Attacker will aim towards a capsule if it is near, an enemy is near, and the attacker is closer to it than the enemy.

While the capsule effect lasts, attacker's foodlimit increases to 10 and will chase ghosts that are near.

While an enemy has capsule effect, defender will runAway from him if in danger.